requirements/base.txt
pymysql
sqlalchemy
protobuf
grpcio
googleapis-common-protos
celery
sentry_sdk
